KRAMER -- prelimiary Multigroup working
Multigroup basic functionality seems to be working, although on initial
testing there appears to be a problem with the planning/execution. It's
not clear at this point where the problem is...hence the weasel words.

Checking in to merge the most current changes, of which I've been told
there are many.