KRAMER -- starting to add service call thread
Adding threaded object to make service calls without the possibility of
hanging the system. Plan is to have a thread running in the panel that
contains a service caller; widgets that want to make a service call then
use a signal to submit the request, with the response being returned at
a later time (without blocking). Copied the service thread from the DRC UI and
started modification...then ran into complexities.

Since the widgets don't share a base class, it seemed like the easiest
way to add a method to process the response would be an interface that
widgets could inherit (mulitply, likely alongside QWidget). This would
also help solve the issue of not having a way to associate a specific widget
with a received signal (in the thread object). Still working through
the kinks...