KRAMER -- change multi-inheritance to superclass
So Qt doesn't play well with multiple inheritance; nice idea to just
implement the service call handling as an interface, but it created an
ambiguous hierarchy (stemming from wanting a signal in the interface,
which would require subclassing QObject).

Anyway, switch to an abstract superclass that extends QWidget, and also
includes the pertinent signal and typedef. Handle the top-level panel
slightly differently, as it extends rviz::Panel (not QWidget); in
addition to actually containing the service thread object, it will also
set up its own version of the signal/method otherwise found in the
ServiceCallWidgetInterface superclass.