KRAMER -- starting on service call thread
Modified ServiceCallThread, converting from 'drc_ui' for use with
'nasa_robot_teleop' package. Not sure if we want to use the heartbeat;
commented out most of related code awaiting decision. Also, looks like
we used signals/slots to return the call results in 'drc_ui', but here
we want to make a method call...that needs to be updated.

In the utils class, moved the method returning the name from an
action_type constant from private to public, as it's used for output
purposes in ServiceCallThread.

Finally, added a static int and built-in naming scheme for the
ServiceCallWidgetInterface superclass; used when registering with the
ServiceCallThread and subsequent result processing.