KRAMER -- removing thread...broken?
Removing service call thread stuff, as it was determined to be
unnecessary. Somehow in doing that, it seems I broke multi-group!

Changes to the multi-group tab do not get carried over to the
individual, active, group tabs. In fact, it doesn't look like the
service call responses are even making it out of the multi-group tab to
the individual group tabs (?! It used to do the right thing !?).

Checking in to move locations...stepping away briefly can only be good.