KRAMER -- prettify, forgot build files
Working on making it prettier, but having difficulty getting the dialog
to resize according to the column widths.

Forgot to update the 'build_config' files with the 'RosparamDialog'
lines (and was wondering why I had to add them again); commit changes
just for those, even though other items aren't complete.