KRAMER -- fixed rosparam consistency issues
The navigation panel was using local storage that wasn't updating when
values on the paramter server changed. An inconsistency was showing up
in the case of 'navigation_mode' changed via the new RosparamDialog;
removed the local variable in 'atlas_path_planner.py', always retreiving
value from the parameter server (which wouldn't be necessary if rospy
had cached parameter functionality).