KRAMER -- tool frame set/clear (not working)
Checking in to preserve code; the showstopper at this point is my
unfamiliarity with python. It's crashing because it doesn't know (or,
rather, I don't know) how to retrieve a constant from a service request
message.

Other code should work, but can't test until I get the service call
right.