KRAMER -- updating service call -> feedback msg
MAKE SURE TO READ LAST PARAGRAPH

The point here is avoid UI lockups/brownouts for long-running tasks,
which occur due to the service call interface that blocks until a task
is complete. The obvious way around this is to put the service calls in
a thread...which would have been a lot of code (not insurmountable, nor
too difficult, but it would have required many changes in many places).

Instead, came up with a (clever) workaround that returns quickly, is
localized to 'interactive_controls.py', and (as a bonus) eliminates
effectively repeated code. The issue stems from the fact that RViz is
C++, but the backing code is python; this leads to two types of
interaction: via interactive markers (set up and run from the python
code), or via RViz panels (embedded in RViz). Since the python node
is separate from RViz, service calls can operate with no UI
interruption; not so for the panels.

More explcitly, interactive markers operate through an 'interactive
marker server' object that uses a pub/sub scheme: the
'InteractiveMarkerUpdate' message relays information from the server
to RViz, while the 'InteractiveMarkerFeedback' message relays
information from RViz to the server. Interactiver marker menu
interactions produce feedback messages that are handled in the
'process_feedback' method, whereas service calls from the panel are
processed in the 'handle_configure' method.

The workaround takes advantage of the 'process_feedback' method by
converting (chosen, long-running) service calls to feedback messages,
publishing them *as if* the came from an RViz interactive marker, then
quickly returning. The callback queue guarantees ordering, while
conversion replaces duplicate code with redirection.

Fixed a few bugs and now it seems to work appropriately for cartesian,
navigation, multi-group, and stored poses. With one caveat: sometimes
an extra click of the 'plan' button is needed. The newly minted
feedback messages are always published (according to 'rostopic echo'
in another terminal), yet sporadically and stubbornly avoid reception.